[ft #161382414] adds article read stats * [ft #161382414] adds read article stats * [ft #161382414] adds article read stats * [ft #161382414] adds article read stats
