+++ .coveragerc
@@ -20,7 +20,17 @@ omit =
     *manage.py
     *serializers.py
     *utils.py
-    *renderers.py
+    *renders.py
     *favorite/apps*
-    
+    *like_articles*
+    *article_filters*
+    *signals*
+    *imports*
+    *share_articles*
+    *test_bookmarks*
+    *exceptions*
+    *test_favorite*
+    *test_profile*
+    *article_views*
+
 show_missing= True
