+++ authors/apps/core/utils/utils.py
@@ -1,6 +1,7 @@
 import jwt
 from rest_framework.exceptions import APIException
 from django.conf import settings
+from django.contrib.auth.models import AnonymousUser
 
 
 class Utils:
